[REFACTOR] 6단계 - 패키지 구조 변경
