+++ src/main/java/roomescape/api/exception/ReservationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.api.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
