+++ src/main/java/roomescape/api/reservations/controller/ReservationsController.java
@@ -1,14 +1,14 @@
-package roomescape.controller.reservations;
+package roomescape.api.reservations.controller;
 
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.dto.request.ReservationRegisterRequest;
-import roomescape.dto.response.ReservationRegisterResponse;
-import roomescape.entity.ReservationEntity;
+import roomescape.api.reservations.dto.request.ReservationRegisterRequest;
+import roomescape.api.reservations.dto.response.ReservationRegisterResponse;
+import roomescape.db.entity.ReservationsEntity;
 import roomescape.enums.ErrorMessage;
-import roomescape.exception.NotFoundException;
+import roomescape.api.exception.NotFoundException;
 
 import java.net.URI;
 import java.util.ArrayList;
@@ -19,11 +19,11 @@ import java.util.concurrent.atomic.AtomicLong;
 @RequestMapping("/reservations")
 public class ReservationsController {
 
-    private List<ReservationEntity> reservationEntities = new ArrayList<>();
+    private List<ReservationsEntity> reservationEntities = new ArrayList<>();
     private AtomicLong index = new AtomicLong();
 
     @GetMapping
-    public ResponseEntity<List<ReservationEntity>> getReservations() {
+    public ResponseEntity<List<ReservationsEntity>> getReservations() {
         return ResponseEntity.status(HttpStatus.OK).body(reservationEntities);
     }
 
@@ -31,22 +31,22 @@ public class ReservationsController {
     public ResponseEntity<ReservationRegisterResponse> createReservation(
             @RequestBody @Valid final ReservationRegisterRequest reservationRegisterRequest
     ) {
-        final ReservationEntity reservationEntity = ReservationEntity.builder()
+        final ReservationsEntity reservationsEntity = ReservationsEntity.builder()
                 .id(index.incrementAndGet())
                 .name(reservationRegisterRequest.name())
                 .date(reservationRegisterRequest.date())
                 .time(reservationRegisterRequest.time())
                 .build();
 
-        reservationEntities.add(reservationEntity);
+        reservationEntities.add(reservationsEntity);
 
         return ResponseEntity.status(HttpStatus.CREATED)
-                .header("Location", URI.create("/reservations/" + reservationEntity.getId()).toString())
+                .header("Location", URI.create("/reservations/" + reservationsEntity.getId()).toString())
                 .body(ReservationRegisterResponse.of(
-                        reservationEntity.getId(),
-                        reservationEntity.getName(),
-                        reservationEntity.getDate(),
-                        reservationEntity.getTime()
+                        reservationsEntity.getId(),
+                        reservationsEntity.getName(),
+                        reservationsEntity.getDate(),
+                        reservationsEntity.getTime()
                 ));
     }
 
@@ -54,16 +54,16 @@ public class ReservationsController {
     public ResponseEntity<Void> deleteReservation(
             @PathVariable(name = "reservationId") final long reservationId
     ) {
-        final ReservationEntity findReservationEntity = reservationEntities.stream()
-                .filter(reservationEntity -> reservationEntity.getId() == reservationId)
+        final ReservationsEntity findReservationsEntity = reservationEntities.stream()
+                .filter(reservationsEntity -> reservationsEntity.getId() == reservationId)
                 .findFirst()
                 .orElse(null);
 
-        if (findReservationEntity == null) {
+        if (findReservationsEntity == null) {
             throw new NotFoundException(ErrorMessage.NOT_FOUND_RESERVATION);
         }
 
-        reservationEntities.remove(findReservationEntity);
+        reservationEntities.remove(findReservationsEntity);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
     }
