+++ src/main/java/roomescape/api/view/controller/RoomescapeViewController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.view;
+package roomescape.api.view.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
